moved tests back to src folder
The compiler is complaining that the tests are not inside the src directory. That is frustrating
