+++ jest.config.js
@@ -1,2 +1,2 @@
 /** @type {import('ts-jest').JestConfigWithTsJest} */
-module.exports = { preset: "ts-jest", testEnvironment: "node", testMatch: [__dirname + "/test/**/*.test.ts"] };
+module.exports = { preset: "ts-jest", testEnvironment: "node", testMatch: [__dirname + "/src/**/test/**/*.test.ts"] };
